Update location of FE tests to reside with components and with .test.js extension
