+++ react_front_end/src/App.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, waitFor } from "@testing-library/react";
-import App from "../../src/App";
+import App from "./App";
 
 test("renders progressbar", async () => {
   const { getByRole } = render(<App />);
